# doc/api.yaml
openapi: 3.0.0
info:
  title: WASAText
  version: 1.0.5
  description: Project made by 2087354. In questa versione ho adattato l'API a seconda del codice Go.

  contact:
    name: Sasha
    url: https://www.youtube.com/watch?v=dQw4w9WgXcQ
    email: bazhura.2087354@studenti.uniroma1.it
...
components:
  securitySchemes:
      bearerAuth:
        type: http
        scheme: bearer
  schemas:
...
    SendMessageRequest:
      type: object
      properties:
        recipientUsername:
          type: string
          description: Username del destinatario diretto. Ignorato se viene fornito un ID di conversazione.
          example: Ciccione
        message:
          $ref: "#/components/schemas/SendMessageContent"
          
      required:
        - message  # `message` deve esistere, non invio roba vuota lol
    SendMessageContent:
      description: Struct inviata in post da SendMessage. Prende i stessi campi di Message ma senza status, author, ID, timestamp e reactions, siccome vengono gestiti dal sistema.
      type: object
      oneOf: # faccio un bordello per fare TEXT OR PHOTO OR TEXT+PHOTO come opzioni.
            - properties:
                Photo:
                  $ref: "#/components/schemas/Photo"
                Text: 
                  type: string
                  example: Ciao! Ti volevo chiedere se mi presti 10$
              required: 
                - Photo
            - properties:
                Photo:
                  $ref: "#/components/schemas/Photo"
                Text:
                  type: string
                  example: Ciao! Ti volevo chiedere se mi presti 10$
              required: 
                - Text
            - properties:
                Photo:
                  $ref: "#/components/schemas/Photo"
                Text:
                  type: string
                  example: Ciao! Ti volevo chiedere se mi presti 10$
              required:
                - Photo
                - Text
    Photo:
      description: A binary file that represent the image.
      type: string
      format: binary
      example: /9j/4AAQ
    UserID: 
      description: Unique identifier/name of a user. Also called Username/Author.
      type: string
      example: "Ciccione"
    GenericID:
      description: A universal multipurpose unique identifier. Should be autogenerated by back-end.
      type: string
      example: "genID"
    User:
      description: A resource representing a user.
      type: object
      properties:
        Username:
          $ref: "#/components/schemas/UserID"
        Photo:
          $ref: "#/components/schemas/Photo"
    Message:
      description: A resource indicating a single message
      type: object
      properties:
        Timestamp:
          description: Timestamp of the message
          type: string
          format: time-stamp
          example: 2025-01-06T12:00:00Z

        Content:
          description: The content of the message, could contain plain text, photo, or both, or none...
          type: object
          properties:
            Photo:
              $ref: "#/components/schemas/Photo"
            Text:
              description: Textual content of the message
              type: string
              example: Ciao! Ti volevo chiedere se mi presti 10$
        MsgID:
          $ref: "#/components/schemas/GenericID"

        Author:
          $ref: "#/components/schemas/UserID"

        Status:
          description: An ENUM representing the checkmarks/status of the message.
          type: string 
          enum:  ['seen','delivered']
          example: 'delivered'
          # TODO setStatus i guess?
        Reactions:
          description: A collection of Reaction(s)
          type: array
          items:
            $ref: "#/components/schemas/Reaction"
...
    Reaction:
      type: object
      properties:
        Author: 
          $ref: "#/components/schemas/UserID" 
        Timestamp:
          description: Timestamp of the reaction
          type: string
          format: time-stamp
          example: 2025-01-06T12:00:00Z
        Emoticon:
          description: An emoticon symbol representing the reaction.
          type: string
          maxLength: 1
          example: 🍆
    ConversationELT:
      description: This is a resource representing a conversation (collection of messages) that will apear in the conversations list. 
      type: object
      properties:
        ConvoID:
          $ref: "#/components/schemas/GenericID"
        DateLastMessage:
          description: Timestamp of the last message
          type: string
          format: time-stamp
          example: 2025-01-06T12:00:00Z
        Preview:
            description: Preview dell'ultimo messaggio. ["📷 Photo" se foto, "📷 TestoPreview..." se foto e testo, "TestoPreview..." se solo testo]
            type: string
            example: "Ciao, ti volevo chiedere se..."
        Messages:
          description: A collection of messages
          #Sorted in reverse chrono order? TODO (i dont know how to implement it and i will think about it later.)
          type: array
          items:
            $ref: '#/components/schemas/Message'
    Group:
      description: Resource representing a Group (A Conversation element and a collection of users.)
      type: object
      properties:
        GroupPhoto:
          $ref: "#/components/schemas/Photo"
        Conversation:
          $ref: "#/components/schemas/ConversationELT"
        Name:
          description: Group name
          type: string
          readOnly: false
        Users:
          description: A collection of users participating in the conversation.
          type: array
          minLength: 2
          # for a group to exist, minimum is 2
          items:
           $ref: '#/components/schemas/User'
...
    Conversations:
      description: Resource representing the list of the user's conversations..
      type: array
      items:
        $ref: '#/components/schemas/ConversationELT'
paths:
...
  /api/users/me/username:
    description: Path utilizzato per l'operazione di setMyUserName
  
    put:  #setMyUserName
      tags: 
        - "Users"
      operationId: setMyUserName
      summary: Aggiorna il nome utente di un utente loggato
      description: Utilizza la stringa passata come parametro per impostare il nuovo nome utente all'utente loggato in questo momento.
      
      security:
        - bearerAuth: []
        
      requestBody:
        required: true
        content:
          text/plain:  
            schema:
              $ref: '#/components/schemas/UserID'
              
      responses:
        "200":
         description: Aggiornato il nome utente con successo.
         content: #ritorna lo user
            application/json:
              schema: 
                $ref: "#/components/schemas/User"
        "404": 
            description: Utente non trovato :( 
        "400":
            description: Input non valido...  
  /api/conversations: 
    description: Conversations implica le conversazioni del CURRENT USER. Utilizzato per l'operazione di getMyConversations.

    get: #getMyConversations
      tags: 
        - "Private Conversation"
        - "Group Conversation"
      description: Recupera le conversazioni dell'utente loggato in questo momento 
      summary: Recupera lista messaggi.
      security:
        - bearerAuth: []
      operationId: getMyConversations
      responses:
        "200":
          description: Hea' ya' go m8, yo' lefty convasations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversations'
...
  /conversations/messages:
    description: Path utilizzato per l'operazione di sendMessage  
  
    post: #sendMessage 
      tags: 
        - "Message" 
      description: Invia il messaggio contenuto nel body sulla conversazione identificata dal parametro.
      summary: Invia un messaggio in una convo data dall'id oppure direttamente ad un utente (se non esiste convo la crea).
      security:
        - bearerAuth: []
      operationId: sendMessage
      
      parameters:
        - name: ID
          in: query
          description: L'id della convo in cui mandare il messaggio
          required: false
          schema:
            $ref: '#/components/schemas/GenericID'
      
      requestBody:
        description: Risorsa del Messaggio da inviare.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      
      responses:
        "200":
          description: All good, message sent bro
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Message"
        "404":
          description: Convo or user not found error.
        "400":
          description: Parametri de schifo mhai dato
      
  #per il forwarding non mi serve la conversazione originaria, solo l'id del messaggio e la convo di destinazione come parametro perciò uso /messages come path, siccome ci sono più operazioni le separo in path (forward/comment ecc.)
...
  /api/groups/members:
    description: Path usato per operazione addToGroup e leaveGroup
  
    put: #addToGroup
      tags: 
        - "Group Conversation" 
      description: Aggiunge l'utente passato nel body al gruppo identificato dal parametro.
    
      security:
        - bearerAuth: []
      operationId: addToGroup
        
      parameters:
        - name: ID
          in: query
          description: L'id della convo-gruppo a cui aggiungere un utente, se vuota => crea nuovo gruppo.
          required: false
          schema:
            $ref: "#/components/schemas/GenericID"
        
      requestBody:
        required: true
        content:
          text/plain:
            schema: 
              $ref: "#/components/schemas/UserID"
        
      responses:
        '200':
          description: All good, user added to group
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Group"  
        '400':
          description: Convo not found error.
  
    delete: #leaveGroup
      tags: 
        - "Group Conversation" 
      description: Elimina l'utente loggato dalla lista degli utenti del gruppo identificato dall'id passato come parametro.
    
      security:
        - bearerAuth: []
      operationId: leaveGroup
      
      parameters:
        - name: ID
          in: query
          description: L'id del gruppo da cui il user corrente (/me) vuole uscire
          required: true
          schema:
            $ref: "#/components/schemas/GenericID"

      responses:
        '204':
          description: All good, you left the group bro
        '400':
          description: Group not found error.
  /api/groups/{ID}/name:
    description: Path usato per operazione setGroupName
    
    put:  #setGroupName
      tags: 
        - "Group Conversation" 
      description: Cambia il nome del gruppo identificato dall'id usando la stringa nel body.
    
      security:
        - bearerAuth: []
      operationId: setGroupName
      
      parameters:
        - name: ID
          in: path
          description: L'id della convo-gruppo a cui cambiare il nome
          required: true
          schema:
            $ref: "#/components/schemas/GenericID"
            
      requestBody:
        required: true
        content:
          text/plain:
            schema: 
              description: Group name string
              type: string
              readOnly: false
      
      responses:
        "200":
         description: Aggiornato il nome gruppo con successo.
         content: #ritorna il gruppo
            application/json:
              schema:  
                $ref: "#/components/schemas/Group"
        "404": 
            description: Gruppo non trovato :(   
  /api/users/me/photo:
    description: Path usato per operazione setMyPhoto
    
    put:  #setMyPhoto
      tags: 
        - "Users"
      description: Imposta la foto profilo dell'utente loggato al momento.
      
      security:
        - bearerAuth: []
      operationId: setMyPhoto
      
      requestBody:
        required: true
        content:
          multipart/form-data: #tecnicamente, se è una richiesta va bene, ho letto sul libro che è meglio non farlo per le risposte :)
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      
      responses:
        "200":
         description: Aggiornato la pfp con successo.
         content: #ritorna il user
            application/json:
              schema: 
                $ref: "#/components/schemas/User"
  /api/groups/{ID}/photo:
    description: Path usato per operazione setGroupPhoto
  
  
    put:  #setGroupPhoto
      tags: 
        - "Group Conversation" 
      description: Imposta la foto profilo del gruppo identificato dal parametro
      security:
        - bearerAuth: []
      operationId: setGroupPhoto
      
      parameters:
        - name: ID
          in: path
          description: L'id della convo-gruppo a cui cambiare la foto
          required: true
          schema:
            $ref: "#/components/schemas/GenericID" 
      
      
      requestBody:
        content:
          multipart/form-data: #tecnicamente, se è una richiesta va bene, ho letto sul libro che è meglio non farlo per le risposte :)
            schema:
              $ref: "#/components/schemas/Photo" 
      
      responses:
        "200":
         description: Aggiornato la pfp con successo.
         content: #ritorna il gruppo
            application/json:
              schema: 
                $ref: "#/components/schemas/Group"   
...